CC-1909: Refactor, move ServiceMap into it's own file
